pkvm: x86: Support IOMMU w/o nested-translation cap
On MTL, the igfx-iommu has no cap of nested-translation support. There is
an intention to let the igfx-iommu work in host iommu identity domain and
leave the inter-vm isolation work to pKVM with the help of using host ept
as the second-stage page translation table. This kind of usage doesn't
rely on hardware nested-translation cap and pKVM needs to support it.

Signed-off-by: Tina Zhang <tina.zhang@intel.com>